+++ vcxproj2cmake.Tests/ConverterTests/ConfigurationTypeTests.cs
@@ -27,7 +27,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(App)
 
                 add_executable(App)
@@ -54,7 +54,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Lib)
 
                 add_library(Lib STATIC)
@@ -80,7 +80,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Dll)
 
                 add_library(Dll SHARED)
@@ -107,7 +107,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(HeaderOnly)
 
                 add_library(HeaderOnly INTERFACE)