+++ vcxproj2cmake.Tests/ConverterTests/LinkLibraryDependenciesTests.cs
@@ -31,14 +31,14 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(Path.Combine("Dll", "CMakeLists.txt"), fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Dll)
 
                 add_library(Dll SHARED)
                 """);
 
             Assert.FileHasContent(Path.Combine("App", "CMakeLists.txt"), fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(App)
 
                 add_executable(App)
@@ -73,14 +73,14 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(Path.Combine("Dll", "CMakeLists.txt"), fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Dll)
 
                 add_library(Dll STATIC)
                 """);
 
             Assert.FileHasContent(Path.Combine("App", "CMakeLists.txt"), fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(App)
 
                 add_executable(App)