+++ vcxproj2cmake.Tests/ConverterTests/ProjectReferenceOrderTests.cs
@@ -41,7 +41,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent("CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Solution)
 
                 add_subdirectory(LibC)
@@ -88,7 +88,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent("CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(Branching)
 
                 add_subdirectory(LibC)