+++ vcxproj2cmake.Tests/ConverterTests/QtTests.cs
@@ -74,7 +74,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project(QtProject)
 
                 find_package(Qt6 REQUIRED COMPONENTS Core Widgets)