+++ vcxproj2cmake.Tests/ConverterTests/QuotingAndEscapingTests.cs
@@ -25,7 +25,7 @@ public partial class ConverterTests
 
             // Assert
             Assert.FileHasContent(@"CMakeLists.txt", fileSystem, """            
-                cmake_minimum_required(VERSION 3.24)
+                cmake_minimum_required(VERSION 4.0)
                 project("Project with \$pecial c#aracters")
 
                 add_executable("Project with \$pecial c#aracters")